moon orbits planet (needs scaling)

+++ Spaceflight/Assets/CelestialBody.cs
@@ -1,4 +1,4 @@
-using System;
+  using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,10 +7,10 @@ public class CelestialBody : MonoBehaviour
 {
     //public Decimal big_mass;
     //public Decimal small_mass;
-    public Decimal gravitational_constant = 6.6743e-11m;
+    public Double gravitational_constant = 6.6743e-11;
     public float gravitational_force;
     public float distance;
-    public Decimal dec_distance;
+    public Double dec_distance;
     public Vector3 angle;
     public Vector3 gravity;
     public float orbital_velocity;
@@ -20,7 +20,7 @@ public class CelestialBody : MonoBehaviour
     public GameObject body1;
     public GameObject body2;
 
-    public Vector3 calculate_gravity(string big_body, Decimal big_mass, string small_body, Decimal small_mass)
+    public Vector3 calculate_gravity(string big_body, Double big_mass, string small_body, Double small_mass)
     {
         body1 = GameObject.Find(big_body);
 
@@ -28,7 +28,7 @@ public class CelestialBody : MonoBehaviour
         
         //6371146 = radius of earth (use this to scale sizes of planets)
         distance = Vector3.Distance(body2.transform.position, body1.transform.position) * 300000;
-        dec_distance = Convert.ToDecimal(distance);
+        dec_distance = Convert.ToDouble(distance);
 
         angle = (body2.transform.position - body1.transform.position);
         gravitational_force = -((float)((gravitational_constant * big_mass * small_mass) / (dec_distance * dec_distance)));
@@ -38,14 +38,14 @@ public class CelestialBody : MonoBehaviour
         return gravity;
     }
 
-    public float get_orbital_velocity(string big_body, Decimal big_mass, string small_body)
+    public float get_orbital_velocity(string big_body, Double big_mass, string small_body)
     {
         body1 = GameObject.Find(big_body);
 
         body2 = GameObject.Find(small_body);
 
         distance = Vector3.Distance(body2.transform.position, body1.transform.position) * 300000;
-        dec_distance = Convert.ToDecimal(distance);
+        dec_distance = Convert.ToDouble(distance);
 
         orbital_velocity = Mathf.Sqrt((float)((gravitational_constant * big_mass) / (dec_distance)));
